Patch Notes
Players will recieve their teams 2 days before the tournament. This is so that teams that have players on an international scale are able to organise practice sessions.

Disconnect rules have been altered to add Early DCs and remove the player dc rule across the whole tournament (since that is extremly hard to track).